Abort if User not authenticated (not found)
The impersonation workflow doesn't check if a user exists or is authenticated before redirecting the request.

This patch will abort the request with a clear error message if the user is not found or, for any reason, the user is not authenticated.

+++ src/Features/UserImpersonation.php
@@ -46,10 +46,11 @@ class UserImpersonation implements Feature
         $ttl = $ttl ?? static::$ttl;
 
         if ($token->created_at->diffInSeconds(Carbon::now()) > $ttl) {
-            abort(403);
+            abort(403, 'Session Expired');
         }
 
         Auth::guard($token->auth_guard)->loginUsingId($token->user_id);
+        abort_unless(Auth::check(), 403, 'User not found');
 
         $token->delete();
